Add files via upload
A func little challenge i encountered during the Python Institute PCAP certification course. The task was to validate a given sudoku input sequence, 
my approach was to extract the fundamental pieces of the input, (the rows, columns , and the 3x3 blocks) and compare each block with its set counterpart. 

Since each of the block's 9 elements have to be unique, we can turn them into a set, eliminating any duplicates, and see if the length is the same 
as for the list counterpart. if it is, the block is valid, and we move on to the next until all blocks validated. 

This by all means was not intended to be the best perforating or the most optimized and expansive program in the world.
but rather a nice practice with list comprehensions, and component abstraction (mainly the blocks inner function.)

Enjoy:))